requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-countries==7.3.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
<<<<<<< HEAD
=======
install==1.3.5
>>>>>>> backend
jsonfield==3.1.0
Pillow==9.0.1
pycodestyle==2.8.0
PyJWT==2.3.0
pyt==1.0.5
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
